Add files via upload
Create a new repository on your Github account and name it Coursera_Capstone. 

If you haven't used Github before, you can find a series of videos to walk you through the process of creating a repository on Github and setting it up and getting started with Git. For Mac users, click here. For Windows users, click here.

Now, start a Jupyter Notebook using any platform that you are comfortable with and do the following:

1. Write some markdown to explain that this notebook will be mainly used for the capstone project.
2. Import the pandas library as pd.
3. Import the Numpy library as np.
4 Print the following the statement: Hello Capstone Project Course!

Push the Notebook to your Github repository and submit a link to the notebook on your Github repository.